feat: improved development setup
I want to have hot-reloading in the documentation when performing
updates on either the tailwindcss plugin or the ui library.

A Justfile was added at the root of the project for
having a better window on the dev setup.

Note: hot-reloading when a linked dependency changes, is done by
tricking Vite to detect a file change and trigger a rebuild, see
`hackwatcher`.